added a confirmation window to the clear button asking if the user wants to delete everything and added icon to the delete button

+++ js/main.js
@@ -53,8 +53,14 @@ $(document).ready(function() {
 
     // Clear all button
     $('.clear-button').click(function() {
-        localStorage.clear();
-        $('.text-area').val("");
+        let confirmation = confirm('Are you sure you want to delete and remove everything from your schedule?');
+        if(confirmation === true) {
+            localStorage.clear();
+            $('.text-area').val("");
+        } else {
+            return;
+        }
+
     })
 
     hourCheck();
